% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mls.R
\name{mls}
\alias{mls}
\title{Build a multline string}
\usage{
mls(...)
}
\arguments{
\item{...}{String fragments to build the string from}
}
\description{
Build a multline string
}
\examples{

mls::mls("a string that is", "composed of bits of strings")
mls::mls("a string that contains", "", "line breaks")
}
